requirements.txt
pandas
duckdb
dagster
streamlit
pytrends
requests
python-dotenv